Fix setting language when no spellchecker available.
Detected when just starting the app, going to compose and nothing worked.

Dont know why but it seems that dictionary is not initialized as fast or
earyl as before.

As a reliable workaround, it changes the spellchecker.dictionary setting,
this way when dictionary is initialized it will have the right language
setted.