Fix wait_for_text to not re-capture screenshot
When using wait_for_text, we try to pass the same image to
Tesseract with different resize ratios to see if it reads
the image better.

During those retries, the messages on screen could have hided yet.
So what we do is memoize the image in the Reader instance and
retry with different resize ratios.