Fix integration tests English change
As thunderbird default language is English, when in a test
I try to set english to trigger the language saving feature,
the logic of the addon says "you already are in english so
nothing changed. Not storing that".

So for the tests, the language change has to be from eng to spa.

User facing that make sense because if you are writing an email
the language of the spellchecker will always be eng, unless the
user change to spa, that the addon will detect that and set spa
on next sessions.
So it is ok as is.