Change TB display font to work better with tesseract
After looking on the internet, there seem to be no agreement
on which font is better (if monospaced, serif, or whatever).

After some tests, seems that the monospaced was the one which
passed the tests.

Formatting extensions tabs with monospace too. Tesseract works
better when all fonts in a page are the same family.