Install spanish dictionary in pre-build profile
After Thunderbid migrated their dictionaries to webExtension, the
script that was installing the extension via xpi file was not working
because webExtensions do not need to have an ID like old xpi had.

As I can't find the way to install a webextension without knowing the
app ID, I've installed the extension in the profile template to have
it installed in all profiles.

I've tried to install the spanish dictionary by making up a random ID
but it did not work.

When I have to migrate the extension to webExtension I'll investigate
this more deeply.