Make extension compatible with thunderbird v66
The profile that works for TB < 60 does not work for TB > 60.
That's why we define two diferent profiles.

I do not know what is the difference but this was the only way.

Probably in the future we can remove the old one and use the one
compatible for v66 for all tests.

Also I tried to install Spanish dictionary using the interactor
(keyboard keys + click on words) but the page that installs
extensions is a web loaded from the internet and tesseract had
problems reading the button that said "Add this to Thunderbird".

Also that meant that any change on that website would affect tests.

Using the preinstalled one we skip this dependency.