Improve wait_for_text implementation
Previous implementation only captured the screen once and then used different
zooms to read the screen.

We could've tried to retry the whole capture screen + read, but it is too slow.
Screen capturing is less than 0.01s but reading the screen at 2X zoom can take 4-6s.
Addon notifications last 3s so 4-6s is too big.

Solution uses ruby threads to make sure we capture the screen the next following
3 seconds, one screen capture every second.

Then the reader loop reads the first screenshot and will skip to the 2nd or 3rd if
the text is not there.

Changes summary:
- Screen is captured 1 time per second during 3 seconds.
- All wait_for_text attempts use 2X zoom. No zoom increase on retries.