Fix heuristics to only count single emails
FreqSuffix was receiving removes when a composed key was deleted from
the LRU set. This is a problem because composed keys are not sent to the
FreqSuffix so we were substracting something that it did not have, so
it was reaching negative values.

A migration should be created to clean user's data.