Addon was not using max size configuraiton
It allowed user to choose a different max size for the LRU structure,
but the value stored was never used in the LRU so all users are
still using 1200 or probably 200, not sure.

This commit does the following:
- When LRU gets serialized, it does not set size any more.
- When we load LRU object, we set max size from config.

This way when we load LRU data from storage, it will still be for old size, but
the first time user updates the LRU (assigns a language to the user),
the size will not be there any more.