Fix LanguageAssigner data persistence between windows
LanguageAssigner instance is shared between windows but first
implementation was keeping a reference of AD.

So the result was that LanguageAssigner was assigning ad.last_lang
and other values to the first AD instance, all the time, and
not to the one that belonged in each case.

We have an AD instance for each window.

Now we provide the AD instance by parameter to #languageChanged.