Fix version comparison logic
Just comparing strings "1.2.3" > "2.3" has a big flaw, that is
that does not work when you add a new figure. "99" < "100" will return
false.

New implementation takes that into account.

For simplicity, compareVersion function does not care on patch-level
string versions. In this addon we just care about Major versions so
it does not matter if its "1.3.2-beta" or any other.