Fixing bug when reinstalling and still there are preferences
Preferences is the old way of storing data and preferences in
Thunderbird. This addon uses local storage, that is scoped to
the extension and gets removed with the extension itself.

When the addon found in prefManager that addon had already migrated
all data to local storage, addon could not find any data, because
local storage was removed on unistallation, but not prefs.