Fix Reader to not create Words with zoomed size
When we use OCR, we scale up images to get better results.

When we receive the words found in the images, we need to calculate
the original size to use the words positions. That calculation was
done only when calculating word center. Now all words have screen
pixel size, so the filter we use to choose Events from right side
of the screen can be without taking into account the upscaling.