CI - Raise error when the searched text appears more than once
During last weeks I've been facing random issues when trying
to open the burguer menu. I was going nuts because I could not
reproduce it.

What was happening is that tesseract does not return the words in the
same order all the time. Don't know what triggers one or other
ordering, but sometintes retruned first the "Events" from the right
side of the screen, and other times the one in the top left menu.

Now, reader returns an error when the needle is found twice. This way
we will know faster which is the root issue.