Fix reply behaviour due to unexpected dictionary changed event
Thunderbird 102 is sending an event of onActiveDictionariesChanged
when initializing the window.

This made the addon think that someone changed the dictionary and assing
the language to the recipients of the reply emal.

As we have no way to know if a human or the window init generated the
event, we just ignore any event generated in the first 500ms of the
window existence.