Fix reply on slow window initialization
As we have no way to distinguish between compose window
init events/changes like the dictionary change, vs
events generated by the user (like a dictionary change),
we ignore any languageChanged event we detect on the early
stages of a compose window (1.5s).

This approach is more a workaround and has its price, so it would be
great to find a better way some day.

Fixes issue #156