Workaround for tesseract glitch that reads the same word twice
Tesseract is inteded to read book pages, not application interfaces.
This means that it has not been trained to do this, so we can expect
it behaves sometimes in a weird way.

In this case we are detecting that repeated words are in the same region
so they are really the same.