chore: increase number of retries on high CPU usage
sometimes 10 retries is not enough to get an execution
that does not get into the CPU loop of the latest beta
releases.

Waiting for a fix to that. Hopefully it will be fixed
once its stable.

Waiting for a fix to that. Hopefully it will be fixed
once its stable.